+++ app/src/renderer/components/wallet/PageWallet.vue
@@ -14,8 +14,7 @@ tm-page(data-title="Wallet", :title="config.devMode ? '' : 'Wallet'")
     tm-list-item(
       :title="wallet.address"
       :btn="'Receive'"
-      :overflow="true"
-      @click.native="copy")
+      :overflow="true")
 
       btn-receive(slot="btn-receive")
 
@@ -41,7 +40,6 @@ tm-page(data-title="Wallet", :title="config.devMode ? '' : 'Wallet'")
 <script>
 import num from "scripts/num"
 import { mapGetters, mapActions } from "vuex"
-import { clipboard } from "electron"
 import { includes, orderBy } from "lodash"
 import Mousetrap from "mousetrap"
 import DataEmptySearch from "common/TmDataEmptySearch"
@@ -119,14 +117,6 @@ export default {
     },
     updateBalances() {
       this.queryWalletState()
-    },
-    copy() {
-      clipboard.writeText(this.wallet.address)
-
-      this.$store.commit(`notify`, {
-        title: `Copied your address to clipboard.`,
-        body: `You can receive Cosmos tokens of any denomination by sharing this address.`
-      })
     }
   },
   mounted() {