+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * speed up of e2e tests removing an expected timeout happening on navigate @faboweb
 * Use unique keys in validator lists. @NodeGuy
 
+### Changed
+
+* updates to how addresses are displayed and how they can be copied @jbibla
+
 ## [0.10.6] - 2018-10-04
 
 ### Fixed